requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml --output-file requirements.txt
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
colorama==0.4.6
    # via tqdm
filelock==3.17.0
    # via
    #   huggingface-hub
    #   transformers
#fsspec==2025.2.0
    # via huggingface-hub
huggingface-hub==0.29.1
    # via
    #   tokenizers
    #   transformers
idna==3.10
    # via requests
joblib==1.4.2
    # via scikit-learn
numpy==2.2.3
    # via
    #   scikit-learn
    #   scipy
    #   transformers
packaging==24.2
    # via
    #   huggingface-hub
    #   transformers
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   transformers
regex==2024.11.6
    # via transformers
requests==2.32.3
    # via
    #   huggingface-hub
    #   transformers
safetensors==0.5.3
    # via transformers
scikit-learn==1.6.1
    # via llm-distillation (pyproject.toml)
scipy==1.15.2
    # via scikit-learn
threadpoolctl==3.5.0
    # via scikit-learn
tokenizers==0.21.0
    # via transformers
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   transformers
transformers==4.49.0
    # via llm-distillation (pyproject.toml)
typing-extensions==4.12.2
    # via huggingface-hub
urllib3==2.3.0
    # via requests
tqdm>=4.67.1
psutil
pynvml
python-dotenv
accelerate>=0.26.0
torch-sparse==0.6.18

black>=25.1.0
dagshub>=0.5.4
datasets>=3.3.0
mlflow>=2.20.2
mypy>=1.15.0
pre-commit>=4.1.0
pylint>=3.3.4
pytest>=8.3.4
ruff>=0.9.6
